Metadata convetions fist attempt
